requirements.txt
# Healthcare Knowledge Graph - Dipendenze
# ====================================

# Framework per l'interfaccia web
streamlit==1.32.0  # Framework per la creazione di applicazioni web interattive

# Analisi dati e grafi
pandas==2.2.0      # Libreria per l'analisi e manipolazione dei dati
numpy==1.26.3      # Libreria per il calcolo numerico
networkx==3.2.1    # Libreria per la creazione e manipolazione di grafi

# Visualizzazione
plotly==5.18.0     # Libreria per la creazione di visualizzazioni interattive
scikit-learn==1.4.0  # Libreria per algoritmi di machine learning (usata per il clustering)

# Gestione dati
python-louvain==0.16  # Libreria per il rilevamento di comunità nei grafi

flask==2.3.3
matplotlib==3.7.2
openai==1.12.0
google-generativeai==0.3.1
python-dotenv==1.0.1
streamlit-plotly-events==0.0.6
requests==2.31.0
spacy==3.7.2


